diabitics-model/requirements.txt
streamlit
pandas
scikit-learn
pillow
matplotlib
